+++ src/stimulus/cli/shuffle_csv.py
@@ -5,7 +5,7 @@ import argparse
 import json
 import os
 
-from stimulus.data.csv import CsvProcessing
+from stimulus.data.data_handlers import CsvProcessing
 from stimulus.utils.launch_utils import get_experiment
